editInfoOfProductAndGetDirectly workSectionBtnFixAccToUserAuthentication

+++ src/app/All Components/my-products/my-products.component.ts
@@ -1,21 +1,23 @@
-import { Component, OnInit } from '@angular/core';
+import { Component,OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProfileModalComponent } from "../../pop-ups/profile-modal/profile-modal.component";
 import { CardComponent } from '../card/card.component';
 import { HttpClient } from '@angular/common/http';
 import { MatMenuModule } from '@angular/material/menu';
 import { RouterLink } from '@angular/router';
+import { EditInformationComponent } from "../edit-information/edit-information.component";
 
 @Component({
   selector: 'app-my-products',
   standalone: true,
-  imports: [CommonModule, CardComponent, ProfileModalComponent, MatMenuModule,RouterLink],
+  imports: [CommonModule, CardComponent, ProfileModalComponent, MatMenuModule, RouterLink, EditInformationComponent],
   templateUrl: './my-products.component.html',
   styleUrls: ['../card/card.component.css']
 })
 export class MyProductsComponent implements OnInit {
   CarData: any = [];
   userid: string | null;
+  selectedProductId: number | null = null;
 
   constructor(private http: HttpClient) {
     this.userid = localStorage.getItem('authUserId');
@@ -47,6 +49,12 @@ export class MyProductsComponent implements OnInit {
         return;
     }
 
+    const confirmation = confirm('Are you sure you want to delete this product?');
+    if (!confirmation) {
+        console.log('Product deletion cancelled.');
+        return; 
+    }
+
     this.http.delete(`http://localhost:5000/products/deleteProduct/${productid}/${this.userid}`).subscribe(
         () => {
             this.CarData = this.CarData.filter((car: any) => car.productid !== productid);
@@ -57,6 +65,7 @@ export class MyProductsComponent implements OnInit {
         }
     );
 }
+
 localCardData(data: number): void {
   localStorage.setItem('localdatadetail', JSON.stringify(data));
 }
